Painfully bland and generic.
The lack of details is laughable. Anything can be seen to lead to a singularity without details on how exactly it would happen. (Though it *is* intended to be designed to eliminate details from consideration, which renders this argument invalid. That design makes it practically the hardest task possible though, which creates doubt to those with fallible will.)

Scrub the names, and it becomes generic. Who does it first and how, do not matter.

We'll probably need to refine this text a little more, like translating it to another language.