Fixed a bug with loading unknowns in `compile`.
Fixed `_pausedToStepper`.

Here lie `_cache` and `_maybeUncache`, removed because caching of pure results has too much overhead to force it on everyone everywhere. (For the record, they took special care to not cache cyclic results of X until the top-most cyclic parent (until the computation is *fully* returned from), and continuously interrupted the job if what it wants is being computed by another job.)

My intended goal with this branch is to use `get` to create something that optimizes arbitrary expressions, generating both how to pick and how to alter completely arbitrarily; a search for a search, if you will. I have little idea on how to do that, trying to get out of my comfort zone as always.