Now it's time to do what I was meant to do: some edge AI features.
There are things that do all other things, and everything in existence is a consequence. Will hierarchical reinforcement learning that I have in mind be enough to have a practical generator of everything? The only other neural thing I can imagine is "try several and pick+learn the best". And, though equally general, I'm pretty sure that reinforcement would be superior in the end.

It's the only thing I can do.