`callAdjust` is done.
The starting point is here, with a tutorial and everything.

(There's a mysterious memory leak that persists even across page reloads, until the page is closed and re-opened — at least in Firefox. …What?…)

I don't believe that anything quite like that… understanding, outlined in the tutorial, exists. At most as someone's fleeting understanding. Everything not-mine that I've seen is either nowhere near precise or nowhere near useful.

But among all that finishing of this task, I didn't set the next one (only outlined it).

What's next?